+++ src/KillerPdf.Engine/Authoring/PdfCheckBoxMark.cs
@@ -1,0 +1,10 @@
+namespace KillerPdf.Engine.Authoring;
+
+public enum PdfCheckBoxMark
+{
+    Check,
+    Cross,
+    Circle,
+    Diamond,
+    Square
+}