+++ src/KillerPdf.Engine/Authoring/PdfPushButtonHighlightMode.cs
@@ -1,0 +1,10 @@
+namespace KillerPdf.Engine.Authoring;
+
+public enum PdfPushButtonHighlightMode
+{
+    None,
+    Invert,
+    Outline,
+    Push,
+    Toggle
+}