+++ src/KillerPdf.Engine/Authoring/PdfTextFieldAlignment.cs
@@ -1,0 +1,8 @@
+namespace KillerPdf.Engine.Authoring;
+
+public enum PdfTextFieldAlignment
+{
+    Left,
+    Center,
+    Right
+}